Split out InteractiveWindowHistoryTests
